+++ integration-tests/worker/CHANGELOG.md
@@ -10,6 +10,12 @@
   - @openfn/engine-multi@1.0.1
   - @openfn/lightning-mock@2.0.1
   - @openfn/ws-worker@1.0.1
+- Updated dependencies [4f5f1dd]
+- Updated dependencies [58e0d11]
+- Updated dependencies [58e0d11]
+  - @openfn/engine-multi@1.1.0
+  - @openfn/ws-worker@1.0.1
+  - @openfn/lightning-mock@2.0.1
 
 ## 1.0.35