+++ packages/cli/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @openfn/cli
 
-## 1.0.1
+## 1.1.1
 
 ### Patch Changes
 
@@ -9,6 +9,14 @@
   - @openfn/compiler@0.0.41
   - @openfn/deploy@0.4.3
   - @openfn/runtime@1.0.1
+## 1.1.0
+
+### Patch Changes
+
+Allow multiple version of the same adaptor to run in the same workflow
+
+- Updated dependencies [4f5f1dd]
+  - @openfn/runtime@1.1.0
 
 ## 1.0.0