+++ packages/lightning-mock/CHANGELOG.md
@@ -10,6 +10,10 @@
   - @openfn/engine-multi@1.0.1
   - @openfn/lexicon@1.0.0
   - @openfn/runtime@1.0.1
+- Updated dependencies [4f5f1dd]
+- Updated dependencies [58e0d11]
+  - @openfn/engine-multi@1.1.0
+  - @openfn/runtime@1.1.0
 
 ## 2.0.0