+++ packages/runtime/CHANGELOG.md
@@ -1,11 +1,16 @@
 # @openfn/runtime
 
-## 1.0.1
+## 1.1.1
 
 ### Patch Changes
 
 - Updated dependencies [2fde0ad]
   - @openfn/logger@1.0.1
+## 1.1.0
+
+### Minor Changes
+
+- 4f5f1dd: Support workflows with different versions of the same adaptor
 
 ## 1.0.0