+++ packages/ws-worker/CHANGELOG.md
@@ -1,6 +1,6 @@
 # ws-worker
 
-## 1.0.1
+## 1.1.1
 
 ### Patch Changes